+++ src/types/types.ts
@@ -1,0 +1,15 @@
+export type Event = {
+    id: string;
+    title: string;
+    date: string;
+    type: 'Workshop' | 'Meeting' | 'Competition' | 'Networking';
+    description: string;
+};
+
+export type SocialLink = {
+    name: string;
+    icon: React.ElementType;
+    url: string;
+    description: string;
+    color: string;
+};
